mysqlsetup

+++ setup.sh
@@ -26,5 +26,16 @@ sudo usermod -aG docker $USER
 sudo chown $USER /var/run/docker.sock
 newgrp docker
 
-docker run -v ./backend/db/init.sh:/init.sh --env-file .env --rm postgres:latest ./init.sh
-docker rmi $(docker images 'postgres' -a -q)
+. ./.env
+
+if [ "$DATABASE_TYPE" = "postgres" ]; then
+    docker run -v ./backend/db/init.sh:/init.sh --env-file .env --rm postgres:latest ./init.sh
+    docker rmi $(docker images 'postgres' -a -q)
+    echo "$DATABASE_TYPE" database init completed;
+fi
+
+if [ "$DATABASE_TYPE" = "mysql" ]; then
+    docker run -v ./backend/db/mysql_init.sh:/mysql_init.sh --env-file .env --rm mysql_init:latest ./mysql_init.sh
+    docker rmi $(docker images 'mysql' -a -q)
+    echo "$DATABASE_TYPE" database init completed";
+fi
